TSAC-5297 -  Automatic Splicing Tool: when there is a hook the rebar rotate
Fix file positions.
